Make Menu not a Modal
