+++ public/index.html
@@ -23,7 +23,5 @@
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
-    <div id="modal-root"></div>
-    <div id="main-menu-root"></div>
   </body>
 </html>
