+++ src/components/App.jsx
@@ -3,7 +3,7 @@ import { Provider } from 'react-redux';
 import MainArea from './MainArea';
 import Header from './Header';
 import CardModalDialog from './CardModalDialog';
-import MenuModalDialog from './MenuModalDialog';
+import MenuPanel from './MenuPanel';
 import { store } from '../store/reducer';
 
 const App = () => {
@@ -12,7 +12,7 @@ const App = () => {
       <Header />
       <MainArea />
       <CardModalDialog />
-      <MenuModalDialog />
+      <MenuPanel />
     </Provider>
   );
 };
